Move MapCategories.js to container component directory
