+++ src/components/container/MapCategories.js
